cli: Add "set node [node] costcenter" command
The valid defined mutable variables on the "node" object are
defined in the API documentation,
https://github.com/IPnett/cloud-BaaS/blob/master/API.md .

This commit extends the capability of the Perl API client to
perform a:
  "perl ipnett-baas.pl set node [node] costcenter [cost center]"
command, to define the costcenter of a certain node.

The cost center must pass validation against a name space that
is given in the user's SAML authentication ticket, and is
user-dependent.